pyproject.toml
[tool.poetry]
name = "picasso-test-task"
version = "0.1.0"
description = ""
authors = ["Denis <panasenkodf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
djangorestframework = "^3.14.0"
celery = "^5.3.6"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
coverage = "^7.4.1"
flake8 = "^7.0.0"
pep8-naming = "^0.13.3"
flake8-broken-line = "^1.0.0"
flake8-return = "^1.2.0"
flake8-isort = "^6.1.1"
flake8-commas = "^2.1.0"
flake8-print = "^5.0.0"
flake8-comments = "^0.1.2"
flake8-docstrings = "^1.7.0"
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
